refactor(hutch): move HtmlPage() from types file to dedicated module
component.types.ts previously contained only type definitions.
HtmlPage() is a runtime function, so it belongs in its own module.
