+++ projects/hutch/src/runtime/web/html-page.ts
@@ -1,0 +1,12 @@
+import type { Component } from "./component.types";
+
+export function HtmlPage(body: string): Component {
+  return {
+    to: (mediaType) => {
+      if (mediaType !== 'text/html') {
+        return { statusCode: 415, body: '' };
+      }
+      return { statusCode: 200, body };
+    },
+  };
+}
